Add retries for txns when we detect a double-spend (#245)
* Add retries for txns when we detect a double-spend

Without this change, we will completely abandon a txn in the event that
we detect a double-spend. Instead, what we should do is wait to see if
the txn eventually gets mined. If it does, then we should pay the user.

The code added by this commit is very ugly, error-prone, and hard to
test. In order to test it, I added temporary code to make the
double-spend check fail with high probability. This triggered the retry
logic, and I verified that the DESO was sent exactly once.

* only enter goroutine if error is double spend (#248)

* only enter goroutine if error is double spend

* flip return values from BlockCypherPushAndWaitForTxn

Co-authored-by: Lazy Nina <81658138+lazynina@users.noreply.github.com>